% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{attrs}
\alias{attrs}
\alias{attrs<-}
\title{Relational data attributes}
\usage{
attrs(x, ...)

attrs(x, ...) <- value
}
\arguments{
\item{x}{a relational schema object, such as a \code{\link{relation_schema}}
or \code{\link{database_schema}} object, or a relational data object, such
as a \code{\link{relation}} or \code{\link{database}} object.}

\item{...}{further arguments passed on to methods.}

\item{value}{A character vector of the same length as \code{attrs(x, ...)}.}
}
\value{
A list, containing a character vector for each element of \code{x}.
}
\description{
Generic function, for fetching attribute sets for elements of a relational
data object.
}
