% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{attrs_class}
\alias{attrs_class}
\title{Relation data attribute default classes}
\usage{
attrs_class(x, ...)
}
\arguments{
\item{x}{an R object, intended to be a schema.}

\item{...}{further arguments passed on to methods.}
}
\value{
A list of character vectors, giving the default class attribute for
each attribute. Attributes are given in the same order as in
\code{\link{attrs_order}}.
}
\description{
Generic function for schemas, giving the default class vectors for
attributes. These are assigned when using \code{\link{create}}, and can be
overridded when using \code{\link{insert}}.
}
