% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalise.r
\name{autonorm}
\alias{autonorm}
\title{Create a normalised entity set from a dataframe}
\usage{
autonorm(df, accuracy, ...)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{accuracy}{a numeric in (0, 1], giving the accuracy threshold threshold
required in order to conclude a dependency.}

\item{...}{further arguments passed on to \code{\link{dfd}}.}
}
\value{
A "database", a list of two elements. See
\code{\link{cross_reference}} for details.
}
\description{
This is a wrapper function for applying \code{\link{dfd}},
\code{\link{flatten}}, \code{\link{normalise}}, \code{\link{decompose}}, and
\code{\link{cross_reference}}, in order.
}
