% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalize-package.r
\docType{package}
\name{autonormalise}
\alias{autonormalise}
\title{autonormalise: Database normalisation for data frames}
\description{
This is a port of Alteryx's AutoNormalizse Python library
(\url{https://github.com/alteryx/autonormalize}), with various changes /
improvements for the R version.
}
