% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalize-package.r
\docType{package}
\name{autonormalize-package}
\alias{autonormalize}
\alias{autonormalize-package}
\title{autonormalize: Database normalization for data frames}
\description{
This is a port of Alteryx's AutoNormalize Python library
(\url{https://github.com/alteryx/autonormalize}), with various changes /
improvements for the R version.
}
\author{
\strong{Maintainer}: Mark Webster \email{markwebster204@yahoo.co.uk}

}
