% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_reference.r
\name{cross_reference}
\alias{cross_reference}
\title{Add foreign key relationships to a normalised database}
\usage{
cross_reference(schema)
}
\arguments{
\item{schema}{a \code{\link{relation_schema}} object, as given by
\code{\link{synthesise}}.}
}
\value{
A \code{\link{database_schema}} object, containing the given relation
schemas and the created foreign key references.
}
\description{
Add foreign key relationships to a normalised database
}
