% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{d2.data.frame}
\alias{d2.data.frame}
\title{Generate D2 input text to plot a data frame}
\usage{
\method{d2}{data.frame}(x, name = NA_character_, nest_level = Inf, ...)
}
\arguments{
\item{x}{a data.frame.}

\item{name}{a character scalar, giving the name of the record, if any. The
name must be non-empty, since it is also used to name the single table in
the plot. Defaults to \code{NA}: if left missing, it is set to "data".}

\item{nest_level}{an integer, giving the amount of nesting allowed when
giving the class of a list column. Since lists can hold anything in R, this
allows showing common element classes and lengths.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scalar character, containing text input for Graphviz.
}
\description{
Produces text input for D2 to make a diagram of a given data frame, usually
rendered with SVG.
}
\details{
The rows in the plotted data frame include information about the attribute
classes.
}
\seealso{
The generic \code{\link{d2}}.
}
