% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.r
\name{decompose}
\alias{decompose}
\title{Decompose a data.frame based on given normalised dependencies}
\usage{
decompose(df, norm_deps)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{norm_deps}{a list of lists, containing normalised dependencies, as for
output from \code{\link{normalise}}.}
}
\value{
a named list of tables, containing the normalised data. Tables are
lists with the following elements: \code{df}, the data.frame containing the
data; \code{keys}, the list of character vectors, representing (candidate)
keys for the table; \code{index}, a character vector, representing the
index / primary key of the table; \code{parents}, containing names of
parent tables, i.e. tables referenced in foreign keys.
}
\description{
Decompose a data.frame based on given normalised dependencies
}
