% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{dependant}
\alias{dependant}
\alias{dependant<-}
\title{Dependants}
\usage{
dependant(x, ...)

dependant(x, ...) <- value
}
\arguments{
\item{x}{an R object. For the given method, a
\code{\link{functional_dependency}}.}

\item{...}{further arguments passed on to methods.}

\item{value}{A character vector of the same length as \code{dependant(x, ...)}.}
}
\value{
A character vector containing dependants.
}
\description{
Generic function, with the only given method fetching dependants for
functional dependencies.
}
