% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivalence.r
\name{df_equiv}
\alias{df_equiv}
\title{Test data frames for equivalence under row reordering}
\usage{
df_equiv(df1, df2)
}
\arguments{
\item{df1, df2}{Data frames.}
}
\value{
A logical.
}
\description{
Test data frames for equivalence under row reordering
}
