% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.r
\name{df_rbind}
\alias{df_rbind}
\title{Combine R Objects by Rows or Columns}
\usage{
df_rbind(...)
}
\arguments{
\item{...}{data frames.}
}
\value{
A data frame containing the \code{...} arguments row-wise.
}
\description{
\code{\link{nrow}} takes "a sequence of vector, matrix or data-frame
arguments", and combines by rows for the latter. However, as of R 4.1,
calling this on data frame with zero columns always returns zero rows, due to
the issue mentioned for \code{\link{df_duplicated}}. This function adds
zero-column data frames as a special case.
}
\seealso{
\code{\link{df_duplicated}}
}
