% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalise.r
\name{find_dependencies}
\alias{find_dependencies}
\title{Finds dependencies within dataframe df with the DFD search algorithm}
\usage{
find_dependencies(df, accuracy, exclude_class = character())
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{accuracy}{a numeric in (0, 1], giving the accuracy threshold threshold
required in order to conclude a dependency.}

\item{exclude_class}{a character vector, indicating classes of attributes to
not consider as members of keys. Attributes are excluded if they inherit
from any given class.}
}
\value{
A list of dependencies.
}
\description{
Finds dependencies within dataframe df with the DFD search algorithm
}
