% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonorm.r
\name{flatten}
\alias{flatten}
\title{Flatten functional dependency list for normalisation}
\usage{
flatten(dependencies)
}
\arguments{
\item{dependencies}{a list, containing functional dependencies as returned by
\code{\link{dfd}}.}
}
\value{
A copy of \code{dependencies}, where the \code{dependencies} element
has been transformed to a list of two-element lists, with the dependency's
determinants in the first element, and its dependent in the second. This is
the format required by \code{\link{normalise}}.
}
\description{
Flatten functional dependency list for normalisation
}
