% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gv.database}
\alias{gv.database}
\title{Generate Graphviz input text to plot databases}
\usage{
\method{gv}{database}(x, name = NA_character_, nest_level = Inf, ...)
}
\arguments{
\item{x}{a database, as returned by \code{\link{autoref}} or
\code{\link{autodb}}.}

\item{name}{a scalar character, giving the name of the database, if any. This
name is used for the resulting graph, to allow for easier combining of
graphs into a single diagram if required.}

\item{nest_level}{an integer, giving the amount of nesting allowed when
giving the class of a list column. Since lists can hold anything in R, this
allows showing common element classes and lengths.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scalar character, containing text input for Graphviz.
}
\description{
Produces text input for Graphviz to make an HTML diagram of a given database.
}
\details{
Each relation in the database is presented as a set of rows, one for each
attribute in the relation. These rows include information about the attribute
classes.
}
\seealso{
The generic \code{\link{gv}}.
}
