% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gv.database_scheme}
\alias{gv.database_scheme}
\title{Generate Graphviz input text to plot database schemes}
\usage{
\method{gv}{database_scheme}(x, name = NA_character_, ...)
}
\arguments{
\item{x}{a database scheme, as given by \code{\link{normalise}} or
\code{\link{cross_reference}}.}

\item{name}{a character scalar, giving the name of the scheme, if any.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scalar character, containing text input for Graphviz or the
\code{DiagrammeR} package.
}
\description{
Produces text input for Graphviz to make an HTML diagram of a given database
scheme.
}
\details{
Each relation in the scheme is presented as a set of rows, one for each
attribute in the relation. These rows do not include information about the
attribute classes.

Any foreign key references are represented by arrows
between the attribute pairs.
}
\seealso{
The generic \code{\link{gv}}.
}
