% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gv.relation_schema}
\alias{gv.relation_schema}
\title{Generate Graphviz input text to plot relation schemas}
\usage{
\method{gv}{relation_schema}(x, name = NA_character_, ...)
}
\arguments{
\item{x}{a relation schema, as given by \code{\link{relation_schema}} or
\code{\link{synthesise}}.}

\item{name}{a character scalar, giving the name of the schema, if any.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scalar character, containing text input for Graphviz.
}
\description{
Produces text input for Graphviz to make an HTML diagram of a given relation
schema.
}
\details{
Each relation in the schema is presented as a set of rows, one for each
attribute in the relation. These rows do not include information about the
attribute classes.
}
\seealso{
The generic \code{\link{gv}}.
}
