% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gv_database_scheme}
\alias{gv_database_scheme}
\title{Generate Graphviz input text to plot database schemes}
\usage{
gv_database_scheme(dbs, dbs_name = NA_character_)
}
\arguments{
\item{dbs}{a database scheme, as given by \code{\link{normalise}} or
\code{\link{cross_reference}}.}

\item{dbs_name}{a character scalar, giving the name of the scheme, if any.}
}
\value{
A scalar character, containing text input for Graphviz or the
\code{DiagrammeR} package.
}
\description{
Produces text input for Graphviz to make an HTML diagram of a given database
scheme.
}
\details{
Each relation in the scheme is presented as a set of rows, one for each
attribute in the relation. These rows have the following contents:
\itemize{
\item On the left, the attribute names.
\item On the right, a depiction of the table's (candidate) keys. Each
column represents a key, and a filled cell indicates that the attribute
in that row is in that key. The keys are given in lexical order, with
precedence given to keys with fewer attributes, and keys with attributes
that appear earlier in the original table's attribute order. Default
output from other package functions will thus have the primary key given
first. In the future, this will be changed to always give the primary key
first.
}

Any foreign key references are represented by arrows
between the attribute pairs.

If \code{dbs_name} is not missing, this name is attached to the resulting
graph in Graphviz. This is to allow easier combination of several such graphs
into a single image, if a user wishes to do so.
}
