% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gv_table}
\alias{gv_table}
\title{Generate Graphviz input text to plot single tables}
\usage{
gv_table(df, df_name)
}
\arguments{
\item{df}{a data.frame, to be plotted as a record.}

\item{df_name}{a character scalar, giving the name of the record.}
}
\value{
A scalar character, containing text input for Graphviz or the
\code{DiagrammeR} package.
}
\description{
Produces text input for Graphviz to make an HTML diagram of a given table,
represented by a data.frame.
}
\details{
The table is presented as a record-like shape, with the following elements:
\itemize{
\item A header with the table's name, and the number of (unique) rows.
\item A set of rows, one for each attribute in the table. These rows have the following contents:
\itemize{
\item On the left, the attribute names.
\item In the middle, a depiction of the table's (candidate) keys. Each
column represents a key, and a filled cell indicates that the attribute
in that row is in that key. The keys are given in lexical order, with
precedence given to keys with fewer attributes, and keys with attributes
that appear earlier in the original table's attribute order. Default
output from other package functions will thus have the primary key given
first. In the future, this will be changed to always give the primary key
first.
\item On the right, the attribute types: specifically, the first element
when passing the attribute's values into \code{\link{class}}.
}
}

The name of the table is also used as the name for the resulting graph, if it
is not missing (\code{NA_character_}). This is to allow easier combination of
several such graphs into a single image, if a user wishes to do so.
}
