% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{insert}
\alias{insert}
\title{Insert data}
\usage{
insert(x, vals, ...)
}
\arguments{
\item{x}{an R object, into which to insert data.}

\item{vals}{a data frame, containing data to insert.}

\item{...}{further arguments pass on to methods.}
}
\value{
an R object of the same class as \code{x}, containing the additional
new data.
}
\description{
Insert data
}
