% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalize.r
\name{normalize.data.frame}
\alias{normalize.data.frame}
\title{Normalise a data.frame based on given dependencies}
\usage{
\method{normalize}{data.frame}(df, dependencies)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{dependencies}{a Dependencies object, giving the dependencies to
determine normalisation with.}
}
\value{
a list of data.frames, containing the normalised data.
}
\description{
Normalise a data.frame based on given dependencies
}
