% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize.Dependencies}
\alias{normalize.Dependencies}
\title{Normalizes dependency relationships}
\usage{
\method{normalize}{Dependencies}(x, df)
}
\arguments{
\item{x}{a Dependencies object, containing the dependencies to be normalised.}

\item{df}{a data.frame, containing the data to be normalised.}
}
\value{
a Dependencies object, containing the normalised dependencies.
}
\description{
Normalizes the dependency relationships in dependencies into new
groups by breaking up all partial and transitive dependencies.
}
