% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalize.r
\name{normalize.entityset}
\alias{normalize.entityset}
\title{Normalise a given entity set}
\usage{
\method{normalize}{entityset}(es, accuracy)
}
\arguments{
\item{es}{an EntitySet object, containing a single data.frame to be
normalised.}

\item{accuracy}{a numeric in (0, 1], giving the accuracy threshold threshold
required in order to conclude a dependency.}
}
\value{
The created EntitySet, containing the tables from normalising the
original data.frame.
}
\description{
Normalise a given entity set
}
