% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.r
\name{normalize}
\alias{normalize}
\title{Normalizes dependency relationships}
\usage{
normalize(dependencies)
}
\arguments{
\item{dependencies}{a list of functional dependencies, each composed of a
list, with an element for the left-hand size and one for the right-hand
side.}
}
\value{
a list of lists. Each such list contains two named elements:
\code{attrs} contains the normalised dependencies, and \code{keys} contains
a list of candidate keys.
}
\description{
Normalizes the dependency relationships in dependencies into new
groups by breaking up all partial and transitive dependencies.
}
