% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_step.r
\name{normalize_dataframe}
\alias{normalize_dataframe}
\title{Normalise a data.frame based on given dependencies}
\usage{
normalize_dataframe(df, dependencies)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{dependencies}{a Dependencies object, giving the dependencies to
determine normalisation with.}
}
\value{
a list of data.frames, containing the normalised data.
}
\description{
Normalise a data.frame based on given dependencies
}
