% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_step.r
\name{normalize_step.data.frame}
\alias{normalize_step.data.frame}
\title{Normalise a data.frame based on given dependencies}
\usage{
\method{normalize_step}{data.frame}(x, dependencies)
}
\arguments{
\item{x}{a data.frame, containing the data to be normalised.}

\item{dependencies}{a Dependencies object, giving the dependencies to
determine normalisation with.}
}
\value{
a list of data.frames, containing the normalised data.
}
\description{
Normalise a data.frame based on given dependencies
}
