% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nudge}
\alias{nudge}
\title{Nudge meta-analysis data}
\format{
A data frame with 447 rows and 25 columns
}
\source{
\url{https://osf.io/fywae/}
}
\usage{
nudge
}
\description{
Nudge meta-analysis data
}
\keyword{datasets}
