% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nudge_database_nonumeric}
\alias{nudge_database_nonumeric}
\title{Example normalisation of Nudge meta-analysis data}
\format{
A database with 154 tables and 793 relationships
}
\usage{
nudge_database_nonumeric
}
\description{
Example normalisation of Nudge meta-analysis data
}
\keyword{datasets}
