% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{parents}
\alias{parents}
\title{Schema parents}
\usage{
parents(x, ...)
}
\arguments{
\item{x}{an R object.}

\item{...}{further arguments passed on to methods.}
}
\value{
A list.
}
\description{
Generic function I intend to remove.
}
