% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_database_scheme}
\alias{plot_database_scheme}
\title{Plot database scheme}
\usage{
plot_database_scheme(
  dbs,
  dbs_name = NA_character_,
  text = FALSE,
  con = stdout()
)
}
\arguments{
\item{dbs}{a database scheme, as given by \code{\link{normalise}} or
\code{\link{cross_reference}}.}

\item{dbs_name}{a character scalar, giving the name of the scheme, if any.}

\item{text}{a logical, indicating whether to write the plot information to a
connection, as a Graphviz input, instead of plotting.}

\item{con}{a connection, to which the Graphviz text input is written if
\code{text} is \code{TRUE}.}
}
\description{
Plot database scheme
}
