% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_table}
\alias{plot_table}
\title{Plot single dataframe}
\usage{
plot_table(df, df_name, text = FALSE, con = stdout())
}
\arguments{
\item{df}{a data.frame, to be plotted as a record.}

\item{df_name}{a character scalar, giving the name of the record.}

\item{text}{a logical, indicating whether to write the plot information to a
connection, as a Graphviz input, instead of plotting.}

\item{con}{a connection, to which the Graphviz text input is written if
\code{text} is \code{TRUE}.}
}
\description{
Plot single dataframe
}
