% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityset.r
\name{plot_tables}
\alias{plot_tables}
\title{Plot dataframes with relationships}
\usage{
plot_tables(x, ..., to_file = FALSE)
}
\arguments{
\item{x}{either a data.frame, or an entity set.}

\item{...}{additional arguments required for the given method.}

\item{to_file}{a logical, indicating whether to write the plot information to
a file, as a Graphviz input, instead of plotting. Currently not
implemented.}
}
\description{
Plot dataframes with relationships
}
