% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{relationships}
\alias{relationships}
\alias{relationships<-}
\title{Schema relationships}
\usage{
relationships(x, ...)

relationships(x) <- value
}
\arguments{
\item{x}{an R object.}

\item{...}{further arguments passed on to methods.}

\item{value}{A list, of the same length as \code{relationships}(x, ...).}
}
\value{
a list, giving relationships.
}
\description{
Generic function, returning present relationships.
}
