% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesise.r
\name{remove_extraneous}
\alias{remove_extraneous}
\title{Remove extraneous components from functional dependencies}
\usage{
remove_extraneous(x, reduce_attributes = TRUE)
}
\arguments{
\item{x}{a \code{\link{functional_dependency}} object.}

\item{reduce_attributes}{a logical, TRUE by default. If TRUE,
\code{dependencies} are checked for determinant attributes that are made
redundant by the other dependencies. This is redundant if
\code{dependencies} is output from \code{discover}, since there will be no
such redundant attributes.}
}
\value{
a minimal cover of the original dependencies, i.e. a subset where
those remaining are not implied by the others. If \code{remove_attributes}
is TRUE, dependencies in the determinant are removed if the other
dependencies show them to be redundant.
}
\description{
Remove extraneous components from functional dependencies
}
\examples{
# a -> b and b -> c imply a -> c
x <- functional_dependency(
  list(list("a", "b"), list("b", "c"), list("a", "c")),
  c("a", "b", "c")
)
remove_extraneous(x)

# a -> b lets us remove b from {a, b} -> c
x <- functional_dependency(
  list(list("a", "b"), list(c("a", "b"), "c")),
  c("a", "b", "c")
)
remove_extraneous(x)
}
