% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{subrelations}
\alias{subrelations}
\title{Database subrelations}
\usage{
subrelations(x, ...)
}
\arguments{
\item{x}{an R object, intended to be some sort of database-like object that
contains relations, such as a \code{\link{database}} object.}

\item{...}{further arguments passed on to methods.}
}
\value{
a relation-type object, or a list of relation-type objects if the
subrelation isn't vectorised.
}
\description{
Generic function, returning subrelations for \code{x}.
}
