% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{add_lookup}
\alias{add_lookup}
\title{Add attribute lookup relations}
\usage{
add_lookup(x, as, ...)
}
\arguments{
\item{x}{a relational schema object, such as a \code{\link{relation_schema}}
or \code{\link{database_schema}} object, or a relational data object, such
as a \code{\link{relation}} or \code{\link{database}} object.}

\item{as}{a character vector of elements from \code{\link{attrs_order}(x)},
indicating which attributes to create lookup tables for.}

\item{...}{further arguments pass on to methods.}
}
\value{
an object of the same class as \code{x}.
}
\description{
Create single-attribute relations for each of a given set of attributes. If
the original object also contains references, then the new lookup relations
are connected to other appearances of their attribute by chains of
references.
}
\details{
Whether an attribute in \code{as} gets a new key relation depends on the
relations already present. An existing relation is considered to be a lookup
if the following hold:
\itemize{
\item The attribute is a simple key for the relation;
\item If \code{x} is a relation data object, then the relation contains all
given values for that attribute.
}

If the original object includes references, then references are added to
connected each attribute's appearances to its lookup. If several relations
could be a lookup for the attribute, then \code{add_lookup} fails due to the
desired result being ambiguous.
}
\examples{
db <- autodb(ChickWeight)
db
add_lookup(db, "Time")
add_lookup(db, "Chick") # Chick is already a key, so this does nothing
# data objects round numeric and complex values before checking given values
add_lookup(db, "weight", digits = 3)
\dontrun{add_lookup(c(db, db), "Chick") # fails: two lookup candidates}
# two lookups without references is fine
rels <- subrelations(db)
add_lookup(c(rels, rels), "Chick")
}
