% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{add_lookup}
\alias{add_lookup}
\title{Add attribute lookup relations}
\usage{
add_lookup(x, as)
}
\arguments{
\item{x}{a relational schema object, such as a \code{\link{relation_schema}}
or \code{\link{database_schema}} object, or a relational data object, such
as a \code{\link{relation}} or \code{\link{database}} object.}

\item{as}{a character vector of elements from \code{\link{attrs_order}(x)},
indicating which attributes to create lookup tables for.}
}
\value{
an object of the same class as \code{x}.
}
\description{
Create single-attribute relations for each of a given set of attributes. If
the original object also contains references, then the new lookup relations are
connected to other appearances of their attribute by chains of references.
}
\details{
If an attribute in \code{as} already appears as a key in a relation, then it
does not have a lookup relation added, and values from other relations are
not added to the relations where it is a key. The exception is if \code{x} is
a relational data object, and no such relation contains all of the given
values for that attribute.
}
\examples{
db <- autodb(ChickWeight)
db
add_lookup(db, "Time")
add_lookup(db, "Chick") # Chick is already a key, so this does nothing
}
