% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalize.r
\name{auto_entityset}
\alias{auto_entityset}
\title{Creates a normalized entity set from a dataframe}
\usage{
auto_entityset(df, accuracy, name = NA)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{accuracy}{a numeric in (0, 1], giving the accuracy threshold threshold
required in order to conclude a dependency.}

\item{name}{a character scalar, giving the name of the created entity set.}
}
\value{
An entity set, containing the data normalised into the required
number of tables.
}
\description{
Creates a normalized entity set from a dataframe
}
