% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalize.r
\name{auto_entityset}
\alias{auto_entityset}
\title{Creates a normalized entity set from a dataframe}
\usage{
auto_entityset(df, accuracy, index = NA, name = NA, time_index = NULL)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{accuracy}{a numeric in (0, 1], giving the accuracy threshold threshold
required in order to conclude a dependency.}

\item{index}{a character scalar, giving the name of the columns intended as
the primary key of \code{df}. Defaults to NA, no preferred primary key.}

\item{name}{a character scalar, giving the name of the created entity set.}

\item{time_index}{a character scalar, giving the name of the time column in
\code{df}.}
}
\value{
An entity set, containing the data normalised into the required
number of tables.
}
\description{
Creates a normalized entity set from a dataframe
}
