% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autodb-package.r
\docType{package}
\name{autodb-package}
\alias{autodb-package}
\title{Database-style normalisation for data.frames}
\description{
Automatic normalisation of a data.frame to third normal form, with the
intention of easing the process of data cleaning. (Usage to design your
actual database for you is not advised.) Originally inspired by Alteryx's
AutoNormalize Python library (https://github.com/alteryx/autonormalize), with
various changes and improvements. Automatic discovery of functional or
approximate dependencies, normalisation based on those, and plotting of the
resulting "database" via Graphviz, with options to exclude some attributes at
discovery time, or remove discovered dependencies at normalisation time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://charnelmouse.github.io/autodb/}
  \item \url{https://github.com/CharnelMouse/autodb}
  \item Report bugs at \url{https://github.com/CharnelMouse/autodb/issues}
}

}
\author{
\strong{Maintainer}: Mark Webster \email{markwebster204@yahoo.co.uk} (\href{https://orcid.org/0000-0002-3351-0686}{ORCID})

Authors:
\itemize{
  \item Mark Webster \email{markwebster204@yahoo.co.uk} (\href{https://orcid.org/0000-0002-3351-0686}{ORCID})
}

}
\keyword{internal}
