% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autokey.r
\name{autokey}
\alias{autokey}
\title{Create a relation from a data frame}
\usage{
autokey(
  df,
  keep_rownames = FALSE,
  digits = getOption("digits"),
  progress = FALSE,
  progress_file = "",
  ...
)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{keep_rownames}{a logical or a string, indicating whether to include the
row names as a column. If a string is given, it is used as the name for the
column, otherwise the column is named "row". Like with the other column
names, the function returns an error if this results in duplicate column
names. Set to FALSE by default.}

\item{digits}{a positive integer, indicating how many significant digits are
to be used for numeric and complex variables. This is used for both
pre-formatting in \code{\link{discover_keys}}, and for rounding the data
before use in \code{\link{insert}}, so that the data satisfies the
resulting schema. A value of \code{NA} results in no rounding. By default,
this uses \code{getOption("digits")}, similarly to \code{\link{format}}.
See the "Floating-point variables" section for \code{\link{discover}} for
why this rounding is necessary for consistent results across different
machines. See the note in \code{\link{print.default}} about \code{digits >=
  16}.}

\item{progress}{a logical, for whether to display progress to the user during
dependency search in \code{\link{discover}}.}

\item{progress_file}{a scalar character or a connection. If \code{progress}
is non-zero, determines where the progress is written to, in the same way
as the \code{file} argument for \code{\link[base]{cat}}.}

\item{...}{further arguments passed on to \code{\link{discover_keys}}.}
}
\value{
A \code{\link{relation}} of length 1, containing the data in
\code{df} and its keys.
}
\description{
This is a wrapper function for applying \code{\link{discover_keys}},
\code{\link{relation_schema}}, \code{\link{create}}, and
\code{\link{insert}}. This takes a data frame and adds its keys, resulting in
a single relation. If only the keys are required, this can be much quicker
than running \code{\link{autodb}}.
}
\examples{
# simple example
autokey(ChickWeight)
}
