% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{create}
\alias{create}
\title{Create instance of a schema}
\usage{
create(x, ...)
}
\arguments{
\item{x}{a relational schema object, representing the schema to create an
instance of, such as a \code{\link{relation_schema}} or
\code{\link{database_schema}} object.}

\item{...}{further arguments passed on to methods.}
}
\value{
An instance of the schema. For example, calling \code{create} on a
\code{\link{database_schema}} creates a \code{\link{database}}, where all
the relations contain zero records.
}
\description{
Create a relation data object, using the given relational schema object, with
the resulting relations empty and ready for data insertion using
\code{\link{insert}}.
}
