% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_reference.r
\name{cross_reference}
\alias{cross_reference}
\title{Add foreign key references to a normalised database}
\usage{
cross_reference(tables)
}
\arguments{
\item{tables}{a relation set, i.e. named list of tables, as given by
\code{\link{decompose}}.}
}
\value{
A "database", defined as a list of two elements:
\itemize{
\item \code{dataframes} contains the given relation set, \code{tables}.
\item \code{relationships} contains a list of length-four character
vectors, describing a single attribute pair in a foreign key
relationship. In order, the elements give the name of the child table,
the attribute in the child table, the parent table, and the attribute in
the parent table. normalised into the required number of tables.
}
}
\description{
Add foreign key references to a normalised database
}
