% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_reference.r
\name{cross_reference}
\alias{cross_reference}
\title{Add foreign key relationships to a normalised database}
\usage{
cross_reference(schema, ensure_lossless = TRUE)
}
\arguments{
\item{schema}{a database schema without relationships, as given by
\code{\link{synthesise}}.}

\item{ensure_lossless}{a logical, TRUE by default. If TRUE, and the
decomposition isn't lossless, an extra relation is added to make the
decomposition lossless.}
}
\value{
A \code{\link{database_schema}} object, containing the given relation
schemas and the created foreign key references.
}
\description{
Add foreign key relationships to a normalised database
}
