% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{d2.relation}
\alias{d2.relation}
\title{Generate D2 input text to plot relations}
\usage{
\method{d2}{relation}(x, name = NA_character_, ...)
}
\arguments{
\item{x}{a \code{\link{relation}}.}

\item{name}{a character scalar, giving the name of the schema, if any.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scalar character, containing text input for D2.
}
\description{
Produces text input for D2 to make a diagram of a given relation, usually
rendered with SVG.
}
\details{
Each relation is presented as a set of rows, one for each
attribute in the relation. These rows include information about the
attribute classes.
}
\seealso{
The generic \code{\link{d2}}.
}
