% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.r
\name{decompose}
\alias{decompose}
\title{Decompose a data.frame based on given normalised dependencies}
\usage{
decompose(df, norm_deps)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{norm_deps}{a list of lists, containing normalised dependencies, as for
output from \code{\link{normalise}}.}
}
\value{
A named list of tables in third normal form, that can reproduce the
original data.frame. This can be considered as a "database": in this
package, we refer to it as a "relation set", and reserve the use of
"database" for when the cross references (foreign keys) are included.

Tables are lists with the following elements:
\itemize{
\item \code{df}, the data.frame containing the data.
\item \code{keys}, the list of character vectors, representing
(candidate) keys for the table.
\item \code{index}, a character vector, representing the index / primary
key of the table.
\item \code{parents}, containing names of parent tables, i.e. tables
referenced in foreign keys.
}
}
\description{
Decomposes a data.frame into several tables, based on the given relations.
It's intended that the relations are derived from a list of functional
dependencies for the same data.frame: using anything else will give undefined
behaviour.
}
