% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{dependent}
\alias{dependent}
\alias{dependent<-}
\title{Dependents}
\usage{
dependent(x, ...)

dependent(x, ...) <- value
}
\arguments{
\item{x}{an R object. For the given method, a
\code{\link{functional_dependency}}.}

\item{...}{further arguments passed on to methods.}

\item{value}{A character vector of the same length as \code{dependent(x, ...)}.}
}
\value{
A character vector containing dependents.
}
\description{
Generic function, with the only given method fetching dependents for
functional dependencies.
}
