% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{detset}
\alias{detset}
\title{Determinant sets}
\usage{
detset(x, ...)
}
\arguments{
\item{x}{an R object. For the given method, a
\code{\link{functional_dependency}}.}

\item{...}{further arguments passed on to methods.}
}
\value{
A list containing determinant sets, each consisting of a character
vector with unique elements.
}
\description{
Generic function, with the only given method fetching determinant sets for
functional dependencies.
}
