% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfd.r
\name{dfd}
\alias{dfd}
\title{DFD algorithm}
\usage{
dfd(df, accuracy, progress = FALSE)
}
\arguments{
\item{df}{a data.frame, the relation to evaluate.}

\item{accuracy}{a numeric in (0, 1]: the accuracy threshold required in order
to conclude a dependency.}

\item{progress}{a logical, for whether to display progress to the user.}
}
\value{
a named list, where the names give the dependent attribute, and each
element is a list of character vectors. Each character vector is a set of
determinant attributes for that dependent attribute.
}
\description{
The DFD algorithm finds all the minimal functional dependencies represented
in a relation/table, represented here in a data.frame. Checks each column to
see if it's unique. If it is unique, it is added as the LHS of a dependency
for every other element. It then loops through all the other non-unique
columns and determines all the LHS that the column depends on. (LHS -->
column)
}
