% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfd.r
\name{dfd}
\alias{dfd}
\title{DFD algorithm
Main loop of DFD algorithm. It returns all the dependencies represented in the
data in dataframe df. Refer to section 3.2 of paper for literature. Checks
each column to see if it's unique. If it is unique, it is added as the LHS of
a dependency for every other element. It then loops through all the other
non-unique columns and determines all the LHS that the column depends on. (LHS
--> column)}
\usage{
dfd(df, accuracy, index = NA, progress = FALSE)
}
\arguments{
\item{df}{a data.frame, the relation to evaluate.}

\item{accuracy}{a numeric in (0, 1]: the accuracy threshold required in order
to conclude a dependency.}

\item{index}{a character, the name of the attribute in the
pre-determined index for the relation. Defaults to NA for no index. This
attribute is marked as unique on the rows, and therefore a determinant of
the other attributes. It is excluded when searching for functional
dependencies.}

\item{progress}{a logical, for whether to display progress to the user.}
}
\value{
a named list, where the names give the dependent attribute, and each
element is a list of character vectors. Each character vector is a set of
determinant attributes for that dependent attribute.
}
\description{
DFD algorithm
Main loop of DFD algorithm. It returns all the dependencies represented in the
data in dataframe df. Refer to section 3.2 of paper for literature. Checks
each column to see if it's unique. If it is unique, it is added as the LHS of
a dependency for every other element. It then loops through all the other
non-unique columns and determines all the LHS that the column depends on. (LHS
--> column)
}
