% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonormalise.r
\name{find_dependencies}
\alias{find_dependencies}
\title{Finds dependencies within dataframe df with the DFD search algorithm}
\usage{
find_dependencies(df, accuracy, ...)
}
\arguments{
\item{df}{a data.frame, containing the data to be normalised.}

\item{accuracy}{a numeric in (0, 1], giving the accuracy threshold threshold
required in order to conclude a dependency.}

\item{...}{further arguments passed on to \code{\link{dfd}}.}
}
\value{
A list of dependencies.
}
\description{
Finds dependencies within dataframe df with the DFD search algorithm
}
