% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_dependency.r
\name{flatten}
\alias{flatten}
\title{Flatten functional dependency list for normalisation}
\usage{
flatten(dependencies)
}
\arguments{
\item{dependencies}{a list, containing functional dependencies as returned by
\code{\link{dfd}}.}
}
\value{
A list of \code{function_dependency} objects, i.e. two-element lists,
with the dependency's determinants in the first element, and its dependent
in the second. Attributes are stored, in the original ordering, in the
\code{attrs} attributes. This is the format required by
\code{\link{normalise}}.
}
\description{
Flatten functional dependency list for normalisation
}
