% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_dependency.r
\name{functional_dependency}
\alias{functional_dependency}
\title{Functional dependency vectors}
\usage{
functional_dependency(FDs, attrs)
}
\arguments{
\item{FDs}{a list of functional dependencies, in the form of two-elements
lists: the first element contains character vector of all attributes in the
determinant set, and the second element contains the single dependent
attribute.}

\item{attrs}{a character vector, giving the names of all attributes. These
need not be present in \code{FDs}, but all attributes in \code{FDs} must be
present in \code{attrs}.}
}
\value{
a \code{functional_dependency} object, containing the list given in
\code{FDs}, with \code{attrs} stored in an attribute of the same name.
Functional dependencies are returned with their determinant sets sorted
according to the attribute order in \code{attrs}.
}
\description{
Creates functional dependencies with length-one dependents.
}
