% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_dependency.r
\name{functional_dependency}
\alias{functional_dependency}
\title{Functional dependency vectors}
\usage{
functional_dependency(FDs, attrs, unique = TRUE)
}
\arguments{
\item{FDs}{a list of functional dependencies, in the form of two-elements
lists: the first element contains character vector of all attributes in the
determinant set, and the second element contains the single dependent
attribute.}

\item{attrs}{a character vector, giving the names of all attributes. These
need not be present in \code{FDs}, but all attributes in \code{FDs} must be
present in \code{attrs}.}

\item{unique}{a logical, TRUE by default, for whether to remove duplicate
dependencies.}
}
\value{
a \code{functional_dependency} object, containing the list given in
\code{FDs}, with \code{attrs} stored in an attribute of the same name.
Functional dependencies are returned with their determinant sets sorted
according to the attribute order in \code{attrs}. Any duplicates found
after sorting are removed.
}
\description{
Creates a set of functional dependencies with length-one dependents.
}
\details{
When several sets of functional dependencies are concatenated, their
\code{attrs} attributes are merged, so as to preserve all of the original
attribute orders, if possible. If this is not possible, because the orderings
disagree, then the returned value of the \code{attrs} attribute is their
union instead.
}
\examples{
fds <- functional_dependency(
  list(list(c("a", "b"), "c"), list(character(), "d")),
  attrs = c("a", "b", "c", "d")
)
detset(fds)
dependent(fds)
attrs(fds)
}
\seealso{
\code{\link{detset}}, \code{\link{dependent}}, and
\code{\link{attrs}} for extracting parts of the information in a
\code{functional_dependency}.
}
