% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gv.data.frame}
\alias{gv.data.frame}
\title{Generate Graphviz input text to plot single tables}
\usage{
\method{gv}{data.frame}(x, name, ...)
}
\arguments{
\item{x}{a data.frame.}

\item{name}{a character scalar, giving the name of the record, if any.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scalar character, containing text input for Graphviz or the
\code{DiagrammeR} package.
}
\description{
Produces text input for Graphviz to make an HTML diagram of a given table,
represented by a data.frame.
}
\details{
The rows in the plotted table include information about the attribute
classes.
}
\seealso{
The generic \code{\link{gv}}.
}
