% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gv.database}
\alias{gv.database}
\title{Generate Graphviz input text to plot databases}
\usage{
\method{gv}{database}(x, ...)
}
\arguments{
\item{x}{a database, as returned by \code{\link{cross_reference}} or
\code{\link{autonorm}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scalar character, containing text input for Graphviz or the
\code{DiagrammeR} package.
}
\description{
Produces text input for Graphviz to make an HTML diagram of a given database.
}
\details{
Each relation in the database is presented as a set of rows, one for each
attribute in the relation. These rows include information about the attribute
classes.
}
\seealso{
The generic \code{\link{gv}}.
}
