% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{gv.relation}
\alias{gv.relation}
\title{Generate Graphviz input text to plot relations}
\usage{
\method{gv}{relation}(x, name = NA_character_, ...)
}
\arguments{
\item{x}{a \code{\link{relation}}.}

\item{name}{a character scalar, giving the name of the schema, if any.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A scalar character, containing text input for Graphviz or the
\code{DiagrammeR} package.
}
\description{
Produces text input for Graphviz to make an HTML diagram of a given relation.
}
\details{
Each relation is presented as a set of rows, one for each
attribute in the relation. These rows include information about the
attribute classes.
}
\seealso{
The generic \code{\link{gv}}.
}
