% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{merge_schemas}
\alias{merge_schemas}
\title{Merge relation schemas in given pairs}
\usage{
merge_schemas(x, to_remove, merge_into, ...)
}
\arguments{
\item{x}{an R object.}

\item{to_remove}{an integer vector, giving the indices for schemas to be
merged into other schemas, then removed.}

\item{merge_into}{an integer vector of the same length as \code{to_remove},
giving the indices for the schemas into which to merge.}

\item{...}{further arguments passed on to methods.}
}
\value{
an R object of the same class as \code{x}, where the relations have
been merged as indicated.
}
\description{
Merge relation schemas in given pairs
}
