% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.r
\name{normalise}
\alias{normalise}
\title{Normalises dependency relationships}
\usage{
normalise(dependencies, progress = 0L, progress_file = "")
}
\arguments{
\item{dependencies}{a list of functional dependencies, as given by
\code{\link{flatten}}: each dependency is a list, contained one character
vector for the left-hand size, and one unit-length character vector for the
right-hand side.}

\item{progress}{an integer, for whether to display progress to the user. 0
(default) displays nothing. 1 notes the start of finding each non-constant
attribute's determinant sets. 2 also briefly describes the status of the
search for an attribute's determinant sets when generating new seeds. 3
also gives the status after visiting each candidate determinant set / node.}

\item{progress_file}{a scalar character or a connection. If \code{progress}
is non-zero, determines where the progress is written to, in the same way
as the \code{file} argument for \code{\link[base]{cat}}.}
}
\value{
A database scheme, represented by a named list of two lists of equal
length, whose elements form pairs. Each pair represents a single relation
scheme in the normalisation:
\itemize{
\item \code{attrs} elements contain the attributes present in the
relation schemes, with attributes in keys given first.
\item \code{keys} elements contain a list of the candidate keys for the
relation schemes.
}
}
\description{
Normalises the dependency relationships in dependencies into a database
scheme, using Bernstein's synthesis.
}
\details{
Bernstein's synthesis is a synthesis algorithm for normalisation of a set of
dependencies into a set of relations that are in third normal form. This
implementation is based on the version given in the referenced paper.

The implementation also includes a common additional step, to ensure that the
resulting decomposition is lossless, i.e. a relation satisfying the given
dependencies can be perfectly reconstructed from the relations given by the
decomposition. This is done by adding an additional relation, containing a
key for all the original attributes, if one is not already present.
}
\references{
Bernstein YEAR, TITLE.
}
