% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.r
\name{normalise}
\alias{normalise}
\title{Normalises dependency relationships}
\usage{
normalise(dependencies)
}
\arguments{
\item{dependencies}{a list of functional dependencies, as given by
\code{\link{flatten}}: each dependency is a list, contained one character
vector for the left-hand size, and one unit-length character vector for the
right-hand side.}
}
\value{
A named list of two lists, with equal length. Each pair represents a
single relation in the normalisation:
\itemize{
\item \code{attrs} elements contain the attributes present, with
attributes in keys given first.
\item \code{keys} elements contain a list of the candidate keys for the
relation.
}
}
\description{
Normalises the dependency relationships in dependencies into a "database", new
groups by breaking up all partial and transitive dependencies.
}
