% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entityset.r
\name{plot_table}
\alias{plot_table}
\title{Plot single dataframe}
\usage{
plot_table(df, df_name, to_file = FALSE)
}
\arguments{
\item{df}{a data.frame, to be plotted as a record.}

\item{df_name}{a character scalar, giving the name of the record.}

\item{to_file}{a logical, indicating whether to write the plot information to
a file, as a Graphviz input, instead of plotting. Currently not
implemented.}
}
\description{
Plot single dataframe
}
