% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot_tables}
\alias{plot_tables}
\title{Plot dataframes with relationships}
\usage{
plot_tables(es, text = FALSE, con = stdout())
}
\arguments{
\item{es}{an entity set, to plotted as linked records.}

\item{text}{a logical, indicating whether to write the plot information to a
connection, as a Graphviz input, instead of plotting.}

\item{con}{a connection, to which the Graphviz text input is written if
\code{text} is \code{TRUE}.}
}
\description{
Plot dataframes with relationships
}
