% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{records}
\alias{records}
\alias{records<-}
\title{Relational data records}
\usage{
records(x, ...)

records(x, ...) <- value
}
\arguments{
\item{x}{an R object.}

\item{...}{further arguments passed on to methods.}

\item{value}{A list of data frames of the same length as \code{records(x,
  ...)}, where each data frame has the same column names as that which it
will replace, in the same order.}
}
\value{
A list containing data frames.
}
\description{
Generic function, for retrieving data contained in a database-like structure.
In particular, this is intended for such structures where the individual
relations can't be accessed with subsetting.
}
