% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{references}
\alias{references}
\alias{references<-}
\title{Schema references}
\usage{
references(x, ...)

references(x) <- value
}
\arguments{
\item{x}{an R object with references, such as a \code{\link{database_schema}}
or \code{\link{database}} object.}

\item{...}{further arguments passed on to methods.}

\item{value}{A list, of the same length as \code{references}(x, ...).}
}
\value{
a list, giving references.
}
\description{
Generic function, returning present (foreign key) references.
}
