% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_class_functions.r
\name{subschemas}
\alias{subschemas}
\title{Schema subschemas}
\usage{
subschemas(x, ...)
}
\arguments{
\item{x}{an R object, intended to be some sort of schema that contains other
schemas, such as a \code{\link{database_schema}} object.}

\item{...}{further arguments passed on to methods.}
}
\value{
A schema-type object, or a list of schema-type objects if the
subschema isn't vectorised. For example, if \code{x} is a
\code{\link{database_schema}}, the result is the contained
\code{\link{relation_schema}}.
}
\description{
Generic function, returning subschemas for \code{x}.
}
